Templating with stacks (week 12)
The assignment provides an opportunity to 
create a stack template using the vector class vector<T> elements. 
Provided functionality include: push(), pop(), top(), clear(), empty();
both pop and top throw a handled exception if called when the stack is empty.
empty is a return of a bool (if empty true else false).